pyproject.toml
[tool.poetry]
name = "datapipe-app"
version = "0.3.2-dev.4"
description = ""
authors = ["Andrey Tatarinov <a@tatarinov.co>"]

readme = "README.md"

packages = [
    { include = "datapipe_app" },
]

include = [
    "datapipe_app/frontend/**/*"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
fastapi = ">=0.69.0"
uvicorn = {extras = ["standard"], version = ">= 0.17.0"}
click = ">=7.1.2"
rich = "^13.3.2"

datapipe-core = {version=">=0.11.12-dev.1, <0.12", allow-prereleases = true}
# datapipe-core = {path="../datapipe"}

prometheus-client = ">= 0.8.0, < 1.0.0"

opentelemetry-instrumentation-fastapi = "0.35b0"

opentelemetry-exporter-gcp-trace = {version="^1.3.0", optional=true}
opentelemetry-exporter-jaeger = {version="^1.8.0", optional=true}

[tool.poetry.extras]
gcp = ["opentelemetry-exporter-gcp-trace"]
jaeger = ["opentelemetry-exporter-jaeger"]

[tool.poetry.plugins."datapipe.cli"]
datapipe_app = "datapipe_app.cli:register_commands"

[tool.poetry.dev-dependencies]
mypy = "*"
sqlalchemy = {extras = ["mypy"], version = "*"}
alembic = "*"
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
